ReadMe File
In this program we attempted to create a profile where the user may enter their information so that it my create an unique and memorable homepage. It needs a lot of work but the idea is that when someone finds the user through the forums, search bar, or on the maps, they may get an idea of who the user is and what they enjoy and can do.