description added
This code shows the backend description of how we can find events in the same city as we are in. 

Certain features include being able to : Find an event on the same day, finding the event in the same city, adding cities and events not in the list, being able to search events by categories of interest. If given time, would add a similar function as add event to be able to add categories people may be interested in. 

This code can also be implemented for people (possible leaders in the same city), and for jobs (the list of jobs available in the same city as you), and much more. 

This code is a basic idea of mapping various events, people and jobs around you to create a culture of connecting people and making it easier for them to share their expertise and talents with each other